Parsing of camera objects is working. Also changed camera width/height to floats.

+++ include/raycast.h
@@ -14,5 +14,5 @@ struct plane {
 };
 
 struct camera {
-    u32 width, height;
+    float width, height;
 };
